[openssl] store specific_config into srv_socket
ssl_ctx is stored into config_storage and it was copied into
the srv_socket.

The link between the config_storage and the srv_socket is hard to
retrieve. This patch stores the specific config into the srv_socket.
ssl_ctx may be retrieve fastly and more informations may be retrieve in
the future.