[openssl] manage tlsext_server_name into connection
tlsext_server_name is a variable used only for SSL connection
but the buffer is allocated for every connections, and it is
useless for simple http connection.

This patch create and free the buffer only for https connections.
This patch allows to keep OpenSSL code into specific functions.

Because tlsext_server_name is always defined some build flags are
useless.