[opensll] anonymize SLL variables into base.h
The SSL variable in the connection object is replaced by a void, because
the type of variable is OpenSSL specific. Another library can use
another structure to define the SSL part of the connection.
The certificats variables is moved into an anonymized structure.

An important modification is that the openssl_patch_connection dump the
pointer of the structure and not the values of the variables. This is
not really dangerous because the variable are specific to the server
socket and not to the connection.