[mod_accesslog] IP address privacy
To respect a user's privacy and make him or her pseudonymous, extra
functionality is hereby added to mod_accesslog.

When enabled using the "accesslog.ip-randomize-interval" configuration
variable, it will randomize a user's IP by concatenating it with a
regularly changing random number
(the randomize interval), hash it (using SHA256), and print 64 bits of
the resulting hash into the log file as a hex string.

This should allow for log file and click stream analysis (continuity of
the resulting hash fingerprint) while respecting the privacy of visitors.

This functionality is only enabled when openssl support is compiled into
lighttpd.