[core] perf: call connection_reset() fewer times
connection_reset() is called at end of request to prepare for keep-alive
or to free resources while waiting for client to close connection.
This is the reason why it is called earlier towards end of request
rather than always waiting until the very end of request when socket
is closed.

connection_reset() used to additionally be called for every new socket
connection, but now is only called when new conns are allocated and
initialized.